docs: Fix `bumpversion` command
The command `bumpversion --new-version 'X.Y.Z'` doesn't work.
It returns the error `the following arguments are required: part`.

The `part` argument doesn't matter when a complete new version number
is specified, but `bumpversion` still complains if that parameter is
missing.

This change fixes the error by adding a dummy `part` argument.

Ref:
- [Slack](https://fynpal-hq.slack.com/archives/CFGURBDBK/p1557152981014100?thread_ts=1556809374.013500&cid=CFGURBDBK)
- [Slack](https://fynpal-hq.slack.com/archives/CFGURBDBK/p1557154032014300?thread_ts=1556809374.013500&cid=CFGURBDBK)
- [Slack](https://fynpal-hq.slack.com/archives/CFGURBDBK/p1557154124014600?thread_ts=1556809374.013500&cid=CFGURBDBK)