data.ref: versioning XML schemas of "factura electrónica"
The contents of the `schemas-xml` directory are separated into
subdirectories, according to the sources used to build the set.
The subdirectories are named considering the source used and the
most recent modification timestamp present between the files in
the set.

Extra:
New enum `cl_sii.base.constants.XmlSchemasVersionEnum` to define
the available XML schema versions