docs: Add creation of release branch to instructions for new releases
The GitHub pull request for the new release uses `develop` as the head
branch. Before the pull request is merged, additional commits may be
pushed to `develop`, and those commits will be merged into the base
branch when the pull request is finally merged, even though what we
wanted to merge was not actually `develop`, but the tag of the new
version.

The documentation is updated with a workaround that "freezes" `develop`
by creating a branch that points to the tag of the new version, so that
the new branch is used as head branch instead of `develop`.