chore(rtc.parse): Disable validation that AEC signature cert is loadable
There is a high number of AEC documents approved by the SII that
cannot be instantiated because the validation that the AEC signature
certificate is loadable fails. For the moment we will disable this
validation so we can instantiate the AEC documents while we look for
a solution to the issue loading the AEC signature certificate.

Ref: https://cordada.aha.io/features/COMPCLDATA-13
Ref: https://github.com/fyntex/lib-cl-sii-python/pull/242#issuecomment-932675576