fix: Python dependency sync check is never executed in CI/CD workflow
The step “Check that compiled Python dependency manifests are up-to-date with their sources”
is supposed to run only for the lowest supported Python version (currently 3.8).
However, commit 751adb045e99946adb1ea447e4344b4281d65900 inadvertently disabled
this step for all Python versions because of an incorrect string comparison
(`3.8.` is not a substring of `3.8`).