fix: Constant RUT_DIGITS_MIN_VALUE should be for the whole range of RUTs
The value `50000000` is the minimum RUT digits for a "Persona Jurídica",
not for the whole range of RUT.

If the minimum RUT digits for a "Persona Jurídica" is desired, use
`PERSONA_JURIDICA_MIN_RUT_DIGITS` instead of `RUT_DIGITS_MIN_VALUE`.

⚠️ Breaking change: The range of possible return values of `Rut.random()`
will be expanded because it uses `RUT_DIGITS_MIN_VALUE`.

Fixes: https://github.com/cordada/lib-cl-sii-python/issues/96