fix(rcv): CSV parser tests for RCV Compras use emisor RUT in context
CSV parser tests for the "Registro de Compras" use the RUT of the "emisor"
in the Marshmallow context, but they should be using the RUT of the "receptor".
Using the RUT of the "emisor" is correct only for the "Registro de Ventas".

This commit fixes the above issue, and also improves the tests so that
this issue would have been detected.